defconfig: use CONFIG_HZ_300=y
Historically, the value has been set to 100 on arm and arm64 platforms.
However, this is an accident of history and is not a good value to use
for interactive devices.

CONFIG_HZ=100 means that a jiffy is 10ms long, which means that
load balancing between CPUs may take 20ms (two jiffies) to happen.
This can significantly contribute to jank on a loaded system.

Recent devices (Nexus 5X, Nexus 6P, Pixel, and Pixel XL)
have shipped with CONFIG_HZ=300. This should have a
negligible power cost while significantly improving runnable times.

Change-Id: Id9806b80873a0e7bc16aa38f480df0ab061ea20c